+++ test/test_go.rb
@@ -2702,6 +2702,20 @@ class TestGoFZF < TestBase
       assert_equal '> 555', lines[index]
     end
   end
+
+  def test_one
+    tmux.send_keys "seq 10 | #{FZF} --bind 'one:preview:echo {} is the only match'", :Enter
+    tmux.send_keys '1'
+    tmux.until do |lines|
+      assert_equal 2, lines.match_count
+      refute lines.any? { _1.include?('only match') }
+    end
+    tmux.send_keys '0'
+    tmux.until do |lines|
+      assert_equal 1, lines.match_count
+      assert lines.any? { _1.include?('only match') }
+    end
+  end
 end
 
 module TestShell